get WalletInfo from kv and get TxMetaData from kv

+++ app/src/main/java/com/platform/entities/WalletInfo.java
@@ -41,7 +41,7 @@ public class WalletInfo {
      */
 
     public int classVersion;
-    public long creationDate;
+    public int creationDate;
     public String name;
     public String currentCurrency;
